Concatenate in chunks to avoid argument list too long
When run via miniwdl slurm/singularity, passing the output of
whatshap phase to bcftools concat directly can result in a very
long command line if there are a large number of regions since
each file is individually mounted as part of the command invocation.
This change breaks the regions into chunks, and each chunk is
concatenated before being passed to the final concatenation command
that produces the fully concatenated output.